Initial LTO commit
Also throw away Krait and Pipe optimizations. Pipe only benefits on HDD,
and Krait tunings seem to spam my log with warnings if I use it with
LTO.

If You are going to use LTO, remember to use gcc-ar, gcc-nm and
gcc-ranlib it's in my big optimizations commit, find it Yourself. Also a
toolchain which supports LTO is needed, for now it's only arter97's
Linaro toolchain fixed by me to use it for ROM.

For now still a lot of modules are disabled, it's really difficult to
find those which work without issues, I'll (slowly) update that list.